chat-bot/llama3_ollama.py
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser
# from langchain_community.llms import Ollama
from langchain_ollama import OllamaLLM

import streamlit as st
import os
from dotenv import load_dotenv
load_dotenv()

os.environ["LANGCHAIN_TRACING_V2"] = "true"
os.environ["LANGCHAIN_API_KEY"] = os.getenv("LANGCHAIN_API_KEY")

# Prompt Template
prompt = ChatPromptTemplate.from_messages(
    [
        ("system", "You are a helpful assistant. Please response to the user queries"),
        ("user", "Question:{question}")

    ]
)

st.title("Langchain Demo with Local Ollama")
input_text = st.text_input("What do you want")

llm = OllamaLLM(model='llama3.2:1b')
output_parser = StrOutputParser()
chain = prompt|llm|output_parser



if input_text:
    st.write(chain.invoke({'question':input_text}))

