chat-bot/requirements.txt
langchain
langchain_openai
langchain_core
langchain_community

streamlit
python-dotenv
langchain-ollama
langserve
fastapi
uvicorn
sse_starlette
requests
langchain_groq